debounce works + fix components & duplicated code & cleaner code
